+++ test/um30xx_4_0_x_decoder.test.js
@@ -400,7 +400,7 @@ describe('UM3xxx tests', () => {
           packet_error_reason: 'value_error',
           packet_fport: 50,
         },errors: [],
-        warnings: ['faulty_downlink_packet: value_error'],
+        warnings: ['faulty_downlink_packet value_error'],
       },
     });
   });