+++ test/utils/pack_test.js
@@ -1,8 +1,7 @@
 import { BinaryPack, BitPack } from '../../src/util/pack';
 
-function err()
-{
-  return {errors:[], warnings:[]};
+function err() {
+  return { errors: [], warnings: [] };
 }
 
 describe('BitPack tests', () => {
@@ -29,20 +28,20 @@ describe('BitPack tests', () => {
   test('addBits overflow', () => {
     var bits = new BitPack(err());
     bits.addBits(3, 2, "key");
-    bits.addBits(22, 7, "key");    
+    bits.addBits(22, 7, "key");
     expect(bits.err.errors.length).toEqual(1);
   });
 
   test('addBit some string', () => {
     var bits = new BitPack(err());
     bits.addBit("true2", "key");
-    expect(bits.err.warnings).toEqual(["key_invalid_boolean_value"]);
+    expect(bits.err.warnings).toEqual(["key invalid_boolean_value_or_key_not_found"]);
   });
 
   test('addBit non-bool integer', () => {
     var bits = new BitPack(err());
     bits.addBit(3, "key");
-    expect(bits.err.warnings).toEqual(["key_invalid_boolean_value"]);
+    expect(bits.err.warnings).toEqual(["key invalid_boolean_value_or_key_not_found"]);
   });
 });
 
@@ -51,20 +50,20 @@ describe('BinaryPack tests', () => {
   test('error too small', () => {
     var pack = new BinaryPack(err());
     pack.addUint8(-1, "key");
-    expect(pack.err.warnings).toEqual(["key_value_too_small"]);
+    expect(pack.err.warnings).toEqual(["key value_too_small"]);
     expect(pack.err.warnings.length).toEqual(1);
   });
 
   test('error too large', () => {
     var pack = new BinaryPack(err());
     pack.addUint8(256, "key");
-    expect(pack.err.warnings).toEqual(["key_value_too_large"]);
+    expect(pack.err.warnings).toEqual(["key value_too_large"]);
   });
 
   test('error not int', () => {
     var pack = new BinaryPack(err());
     pack.addUint8("abc", "key");
-    expect(pack.err.warnings).toEqual(["key_value_not_integer"]);
+    expect(pack.err.warnings).toEqual(["key value_not_integer"]);
   });
 
   test('encode int8', () => {
@@ -104,7 +103,7 @@ describe('BinaryPack tests', () => {
     pack.addFloat(1, "key");
     expect(pack.buffer).toEqual([0x00, 0x00, 0x80, 0x3F]);
     expect(pack.err.warnings.length).toEqual(0);
-  });  
+  });
 
   test('encode float -1', () => {
     var pack = new BinaryPack(err());